complete Process Object

+++ Control.hpp
@@ -25,12 +25,12 @@ public:
         
     }
     void test_direction();
-    void Control_left(double value);
-    void Control_right(double value);
-    void Control_ahead(double value);
-    void Control_Back(double value);
-    void Control_Turn_back(double value);
-    string byte2_4byte(double value);
+    void Control_left(int value);
+    void Control_right(int value);
+    void Control_ahead(int value);
+    void Control_Back(int value);
+    void Control_Turn_back(int value);
+    string byte2_4byte(int value);
     /*input one char without "\n"*/
     int getch(void);
 private:
